+++ examples/standalone-submodule-with/default.nix
@@ -40,7 +40,7 @@ in
 
   config = {
     conan = {
-      profiles = {
+      profiles.default = {
         settings.compiler."compiler.cppstd" = "14";
         settings._.build_type = "Debug";
       };